Tue, Sep 26, 2023, 12:15 PM -04:00

+++ chapter1/cash/cash.c
@@ -14,15 +14,15 @@ int main(void)
        // C is Change inputed by user
        int helper = (C * 100);
        // helper is a variable which turns decimal inputs into a whole number
-       float Q = (helper / 25);
+       int Q = (helper / 25);
        // Q is Max Quarters needed
        int R1 = helper % 25;
        // R1 is the remainder of change in cents still needed after we've taken out all of our Quarters
-       float D = (R1 / 10);
+       int D = (R1 / 10);
        // D is Max Dimes needed
        int R2 = (R1 % 10);
        // R2 is the remainder of change in cents still needed after we've taken out all of our Quarters AND Dimes
-       float N = (R2 / 5);
+       int N = (R2 / 5);
        // N is Max Nickles needed
        int P = (R2 % 5);
        // P is the remainder of change in cents still needed after we've taken out all of our Quarters AND Dimes AND Nickles leaving just Pennies
